% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mopex_data_download.R
\name{data_download}
\alias{data_download}
\title{data dowload from mopex dataset}
\usage{
data_download(
  id,
  url = "https://hydrology.nws.noaa.gov/pub/gcip/mopex/US_Data/Us_438_Daily/",
  save = FALSE,
  path = NULL,
  file_name = NULL
)
}
\arguments{
\item{id}{catchment id}

\item{url}{database url}

\item{save}{logical - save file to local disk}

\item{path}{path of the file}

\item{file_name}{filename}
}
\description{
data dowload from mopex dataset
}
\examples{


}
