% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mopex_data_download.R
\name{get_id}
\alias{get_id}
\title{Get IDs from mopex database}
\usage{
get_id(
  url = "https://hydrology.nws.noaa.gov/pub/gcip/mopex/US_Data/Us_438_Daily/"
)
}
\arguments{
\item{url}{database url}
}
\description{
Get IDs from mopex database
}
\examples{

get_id()

}
