% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot.hello}
\alias{plot.hello}
\title{plot method for hello class}
\usage{
\method{plot}{hello}(x, ...)
}
\arguments{
\item{x}{hello thingy}

\item{...}{other arguments}
}
\description{
plot method for hello class
}
\examples{

plot(x = hello())

}
